+++ express-zod-api/src/deep-checks.ts
@@ -1,12 +1,12 @@
 import * as R from "ramda";
-import { globalRegistry, z } from "zod";
+import { z } from "zod";
 import { ezBufferBrand } from "./buffer-schema";
 import { ezDateInBrand } from "./date-in-schema";
 import { ezDateOutBrand } from "./date-out-schema";
 import { DeepCheckError } from "./errors";
 import { ezFormBrand } from "./form-schema";
 import type { IOSchema } from "./io-schema";
-import { brandProperty } from "./brand";
+import { getBrand } from "./brand";
 import type { FirstPartyKind } from "./schema-walker";
 import { ezUploadBrand } from "./upload-schema";
 import { ezRawBrand } from "./raw-schema";
@@ -53,7 +53,7 @@ export const hasCycle = (
 export const findRequestTypeDefiningSchema = (subject: IOSchema) =>
   findNestedSchema(subject, {
     condition: (schema) => {
-      const { [brandProperty]: brand } = globalRegistry.get(schema) || {};
+      const brand = getBrand(schema);
       return (
         typeof brand === "symbol" &&
         [ezUploadBrand, ezRawBrand, ezFormBrand].includes(brand)
@@ -81,7 +81,7 @@ export const findJsonIncompatible = (
   findNestedSchema(subject, {
     io,
     condition: (zodSchema) => {
-      const { [brandProperty]: brand } = globalRegistry.get(zodSchema) || {};
+      const brand = getBrand(zodSchema);
       const { type } = zodSchema._zod.def;
       if (unsupported.includes(type)) return true;
       if (brand === ezBufferBrand) return true;